requirements.txt
apispec==6.3.0
attrs==23.1.0
autopep8==2.0.2
bcrypt==4.0.1
black==23.7.0
blinker==1.6.2
click==8.1.6
coverage==7.2.7
DateTime==5.2
flake8==6.1.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-rest-api==0.17.1
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
greenlet==2.0.2
gunicorn==21.2.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mistune==3.0.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
uuid==1.30
webargs==8.3.0
Werkzeug==2.3.6
zope.interface==6.0
